requirements.txt

streamlit
pandas
openpyxl
sqlalchemy
python-dotenv